requirements.txt
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
